Move refactor classes based on respective modules
With Java 9 and Project Jigsaw, to replace a class, it requires that we explicitly specify the name of the module that it belongs, like:

    javac --patch-module java.base=src -d mypatches/java.base \
            src/java.base/java/util/concurrent/ConcurrentHashMap.java

Or it will show errors, like:
    [javac] error: package exists in another module: java.base
    [javac] package java.io;

i.e: Java 9 does not allow split packages in different modules (packages having the same name exist in different modules)
Move refactor these split packages into seperate directories to make patching easier.

Since we got sources for multiple modules in the same tree, move them to directories based on their respective modules, for ease of
compilation.

References:
http://openjdk.java.net/projects/jigsaw/quick-start
https://docs.oracle.com/javase/9/docs/api/module-overview-frame.html